refactor: 관리자 api 수정
