+++ marketplace/src/main/java/store/carjava/marketplace/app/marketplace_car/controller/MarketplaceCarController.java
@@ -114,37 +114,6 @@ public class MarketplaceCarController {
         return ResponseEntity.ok(cars);
     }
 
-    @Operation(summary = "차량 판매 승인", description = "특정 차량의 판매를 승인합니다.")
-    @PutMapping("/car/sale/approve")
-    public ResponseEntity<String> approveCar(
-            @RequestParam(name = "carId") String carId,
-            @RequestParam(name = "testDriveCenterId") Long testDriveCenterId,
-            @RequestParam(name = "price") Long price,
-            @RequestParam(name = "exteriorImages", required = false) List<MultipartFile> exteriorImages,
-            @RequestParam(name = "interiorImages", required = false) List<MultipartFile> interiorImages,
-            @RequestParam(name = "wheelImages", required = false) List<MultipartFile> wheelImages,
-            @RequestParam(name = "additionalImages", required = false) List<MultipartFile> additionalImages
-    ) {
-        try {
-            // 모든 이미지 파일들을 하나의 리스트로 통합
-            List<MultipartFile> allImages = new ArrayList<>();
-            if (exteriorImages != null) allImages.addAll(exteriorImages);
-            if (interiorImages != null) allImages.addAll(interiorImages);
-            if (wheelImages != null) allImages.addAll(wheelImages);
-            if (additionalImages != null) allImages.addAll(additionalImages);
-
-
-
-            marketplaceCarService.approveCar(carId, testDriveCenterId, price, allImages);
-            return ResponseEntity.ok("차량 판매가 승인되었습니다.");
-        } catch (IllegalArgumentException e) {
-            return ResponseEntity.badRequest().body(e.getMessage());
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body("서버 오류가 발생했습니다. 나중에 다시 시도해주세요.");
-        }
-    }
-
     @Operation(summary = "최종 차량 판매 완료", description = "차량의 최종 판매를 완료합니다.")
     @PutMapping("/car/sale/complete-sale")
     public ResponseEntity<String> completeSaleCar(
