+++ marketplace/src/main/resources/templates/admin/car-registration-modal.html
@@ -195,7 +195,7 @@
         });
 
         try {
-          const response = await fetch('/car/sale/approve', {
+          const response = await fetch('/car-sales/approve', {
             method: 'PUT',
             body: formData
           });
@@ -256,7 +256,7 @@
           <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
         </div>
         <div class="modal-body">
-          <form id="carDetailForm" action="api/car/sale/approve" method="post" enctype="multipart/form-data">
+          <form id="carDetailForm" action="/car-sales/approve" method="post" enctype="multipart/form-data">
             <!-- PUT 메서드를 처리하기 위한 hidden 필드 -->
             <input type="hidden" name="_method" value="put">
